changed all unnecessary atol() calls to atoi()
git-svn-id: file:///svn/uhexen2/code/trunk@2132 eaf649e6-6c13-0410-b194-84221022858e

+++ hexenworld/Server/pr_edict.c
@@ -928,7 +928,7 @@ char *ED_ParseEdict (char *data, edict_t *ent)
 		}
 		else if (Q_strcasecmp(keyname,"CD") == 0)
 		{
-			sv.cd_track = (byte)atol(com_token);
+			sv.cd_track = (byte)atoi(com_token);
 			continue;
 		}
 
@@ -1135,7 +1135,7 @@ void PR_LoadProgs (void)
 		// Line #1 : <number of lines excluding this one>
 		// Line #2+: <map name><one space><prog filename>
 		fgets(build, sizeof(build), FH);
-		j = atol(build);
+		j = atoi(build);
 		for (i = 0; i < j; i++)
 		{
 			k = 0;
